% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_measures.R
\name{angular_mean}
\alias{angular_mean}
\title{Angular mean}
\usage{
angular_mean(x, period = 2 * pi, revolution = 0)
}
\arguments{
\item{x}{Numeric vector with circular data.}

\item{period}{A numeric value declaring the period of the circular data. The
default is \code{2 * pi}.}

\item{revolution}{A numeric value indicating the number of revolutions to
add to the circular mean. The default is \code{revolution = 0}, which keeps the
circular mean in the standard interval \verb{[0, period[}. If \code{revolution = NULL},
the mean revolution is added.}
}
\value{
The average numeric value of the circular data.
}
\description{
Circular data repeat every period which has to be taken into account
for calculating the averages.
}
\details{
After transformation of the data to radians on the unit circle, the angular
arithmetic mean can be calculated as follows:

\deqn{\bar{\theta} = \text{atan2}\left(\frac{1}{n}\sum_{i=1}^n\text{sin}
(\theta_i),\frac{1}{n}\sum_{i=1}^n\text{cos}(\theta_i)\right)}

Because of the periodicity of circular data, the circular mean can be offset
by \eqn{n} revolutions such that \eqn{\bar{\theta} = \bar{\theta} 
+ 2\pi n}. This can be controlled with the parameter \code{revolution}.
}
\seealso{
Other location_measures: 
\code{\link{circular_location}},
\code{\link{weighted_angular_mean}()}
}
\concept{location_measures}
