% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_measures.R
\name{angular_sd}
\alias{angular_sd}
\title{Angular standard deviation}
\usage{
angular_sd(x, period = 2 * pi)
}
\arguments{
\item{x}{Numeric vector with circular data.}

\item{period}{A numeric value declaring the period of the circular data. The
default is \code{2 * pi}.}
}
\value{
The angular standard deviation of the circular data as numeric value.
}
\description{
Circular data repeat every period which has to be taken into account
for calculating the standard deviation around the mean angle.
}
\details{
After transformation of the circular data to radians on the unit circle, the
circular data can be expressed as complex vectors \eqn{z = e^{i\theta\cdot 
n}}. If the circular data distribution is a wrapped normal or von Mises
distribution, the average value of \eqn{z} (the mean resultant) is given by
\eqn{\langle z\rangle = e^{i\mu - \frac{\sigma^2}{2}}}, with the mean angle
\eqn{\mu}. The length of the mean resultant is given by \eqn{|\langle z
\rangle| = e^{- \frac{\sigma^2}{2}} = \rho}. This yields an expression for
the angular standard deviation \eqn{\sigma} around \eqn{\mu}:
\deqn{\sigma = \sqrt{\text{ln}(\rho^{-2})} = \sqrt{-2\text{ln}(\rho)}}
This is used to calculate the angular standard deviation. The value of
\eqn{\rho} is given by the following expression:
\deqn{|\rho| = \sqrt{\bar{z}\bar{z^*}} = \sqrt{\left(\frac{1}{n}\sum_{i=1}^n
\text{sin}(\theta_i)\right)^2 + \left(\frac{1}{n}\sum_{i=1}^n\text{cos}
(\theta_i)\right)^2}}
}
\seealso{
Other spread_measures: 
\code{\link{angular_sd_yamartino}()},
\code{\link{circular_spread}}
}
\concept{spread_measures}
