% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_measures.R
\name{circular_location}
\alias{circular_location}
\title{Circular measures of location}
\description{
Circular data repeat after every period, which is challenging for
calculating location (mean) and dispersion measures (standard deviation).
}
\details{
This package implements the following measures of location for circular data.
}
\seealso{
Other location_measures: 
\code{\link{angular_mean}()},
\code{\link{weighted_angular_mean}()}
}
\concept{location_measures}
