% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_measures.R
\name{circular_spread}
\alias{circular_spread}
\title{Circular measures of spread}
\description{
Circular data repeat after every period, which is challenging for
calculating location (mean) and dispersion measures (standard deviation).
}
\details{
This package implements the following measures of spread for circular data.
}
\seealso{
Other spread_measures: 
\code{\link{angular_sd}()}
}
\concept{spread_measures}
