% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional_speed.R
\name{directional_speed}
\alias{directional_speed}
\title{Directional speed}
\description{
Some directional data in 2-dimensions include velocities (speed measures),
e.g. wind directions or the intracellular movement of mitochondria.
This package implements the following measures of directional speed. Many of
them are based on Yamartino 1984.
}
\seealso{
Other directional_speed: 
\code{\link{angular_sd_yamartino}()},
\code{\link{weighted_angular_mean}()}
}
\concept{directional_speed}
