% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-density-circular.R,
%   R/stat-density-circular.R
\name{geom_density_circular}
\alias{geom_density_circular}
\alias{stat_density_circular}
\title{Smoothed circular density estimates}
\usage{
geom_density_circular(
  mapping = NULL,
  data = NULL,
  stat = "density_circular",
  position = "identity",
  ...,
  period = 2 * pi,
  bw = NULL,
  adjust = 1,
  kernel = "vonmises",
  n = 512,
  K = NULL,
  trim = FALSE,
  bounds = c(-Inf, Inf),
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_density_circular(
  mapping = NULL,
  data = NULL,
  geom = "area",
  position = "stack",
  ...,
  period = 2 * pi,
  bw = NULL,
  adjust = 1,
  kernel = "vonmises",
  n = 512,
  K = NULL,
  trim = FALSE,
  na.rm = FALSE,
  bounds = c(-Inf, Inf),
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{period}{The period of the cyclical data. A numeric value with default
of \code{2 * pi} radians.}

\item{bw}{The smoothing bandwidth to be used.
If numeric, the concentration parameter of the circular smoothing kernel.
If character, a rule to choose the bandwidth. The following bandwidth
selectors are available:
\itemize{
\item \code{bw = "log"} estimates the smoothing bandwidth as \code{2^log10(n)}. This is
not optimal for small \code{n} but quick for large \code{n}.
\item \code{bw = "taylor"} estimates the smoothing bandwidth with \code{bw_taylor()}.
\item \code{bw ="nrd.trigmoments"} estimates the smoothing bandwidth with
\code{\link[circular:bw.circular]{circular::bw.nrd.circular()}} and uses \code{trigmoments} to compute the
von Mises concentration parameter.
\item \code{bw = "nrd.ml"} estimates the smoothing bandwidth with
\code{\link[circular:bw.circular]{circular::bw.nrd.circular()}} and uses \code{ML} to compute the von Mises
concentration parameter.
\item \code{bw = "ml"} estimates the smoothing bandwidth with
\code{\link[circular:bw.circular]{circular::bw.cv.ml.circular()}}.
\item \code{bw = "mse"} estimates the smoothing bandwidth with
\code{\link[circular:bw.circular]{circular::bw.cv.mse.circular()}}.
Currently, wrapped normal distribution kernels are not working because they
require that x is within (0,1).
Note that automatic calculation of the bandwidth does not take weights
into account.
}}

\item{adjust}{A multiplicate bandwidth adjustment. This makes it possible
to adjust the bandwidth while still using the a bandwidth estimator.
For example, \code{adjust = 1/2} means use half of the default bandwidth.}

\item{kernel}{Kernel. For circular data, the kernel estimation uses the von
Mises distribution implemented in \code{\link[circular:density.circular]{circular::density.circular()}}. A
character string with value \code{kernel = "vonmises"} by default.}

\item{n}{number of equally spaced points at which the density is to be
estimated, should be a power of two. Default is \code{n = 512}).}

\item{K}{The number of terms used to calculate the wrapped normal
distribution. This is currently not working.}

\item{trim}{If \code{FALSE}, the default, each density is computed on the
full range of the data. If \code{TRUE}, each density is computed over the
range of that group: this typically means the estimated x values will
not line-up, and hence you won't be able to stack density values.
This parameter only matters if you are displaying multiple densities in
one plot or if you are manually adjusting the scale limits.}

\item{bounds}{Known lower and upper bounds for estimated data. Default
\code{c(-Inf, Inf)} means that there are no (finite) bounds. If any bound is
finite, boundary effect of default density estimation will be corrected by
reflecting tails outside \code{bounds} around their closest edge. Data points
outside of bounds are removed with a warning.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{orientation}{The orientation of the layer. The default (\code{NA})
automatically determines the orientation from the aesthetic mapping. In the
rare event that this fails it can be given explicitly by setting \code{orientation}
to either \code{"x"} or \code{"y"}. See the \emph{Orientation} section for more detail.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{geom, stat}{Use to override the default connection between
\code{geom_density_circular()} and \code{stat_density_circular()}.}
}
\description{
Computes and draws circular kernel density estimate. This is a useful
alternative to regular kernel density estimate, which fail to capture
cyclical (circular, periodic, directional, polar) data properly.
}
\section{Orientation}{

This geom treats each axis differently and, thus, can thus have two orientations. Often the orientation is easy to deduce from a combination of the given mappings and the types of positional scales in use. Thus, ggplot2 will by default try to guess which orientation the layer should have. Under rare circumstances, the orientation is ambiguous and guessing may fail. In that case the orientation can be specified directly using the \code{orientation} parameter, which can be either \code{"x"} or \code{"y"}. The value gives the axis that the geom should run along, \code{"x"} being the default orientation you would expect for the geom.
}

\section{Aesthetics}{

\code{geom_density()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \code{alpha}
\item \code{colour}
\item \code{fill}
\item \code{group}
\item \code{linetype}
\item \code{linewidth}
\item \code{weight}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}

\section{Computed variables}{

These are calculated by the 'stat' part of layers and can be accessed with \link[=aes_eval]{delayed evaluation}.
\itemize{
\item \code{after_stat(density)}\cr circular density estimate.
\item \code{after_stat(count)}\cr density * number of points - useful for stacked density plots.
\item \code{after_stat(scaled)}\cr density estimate, scaled to maximum of 1.
\item \code{after_stat(n)}\cr number of points.
\item \code{after_stat(ndensity)}\cr alias for \code{scaled}, to mirror the syntax of \code{\link[=stat_bin]{stat_bin()}}.
}
}

\seealso{
\code{\link[circular:density.circular]{circular::density.circular()}}

\code{\link[=bw_taylor]{bw_taylor()}}

\code{\link[circular:bw.circular]{circular::bw.nrd.circular()}}

\code{\link[circular:bw.circular]{circular::bw.cv.ml.circular()}}

\code{\link[circular:bw.circular]{circular::bw.cv.mse.circular()}}
}
