% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-density-circular.R, R/geom-mode-point.R,
%   R/stat-density-circular.R, R/stat-mode-circular.R, R/stat-mode.R,
%   R/ggproto-classes.R
\docType{data}
\name{GeomDensityCircular}
\alias{GeomDensityCircular}
\alias{GeomModepoint}
\alias{StatDensityCircular}
\alias{StatModeCircular}
\alias{StatMode}
\alias{ggcyclical-extensions}
\title{ggcyclical extensions to ggplot2}
\description{
This manual page lists all the exported ggproto classes defined in
\code{ggcyclical}, which were used to construct \verb{stat_*} and \verb{geom_*} functions.
These extend the \code{ggplot2} base functionality in accordance
with its reference manual.
The \verb{geom_*} functions provide new visualization marks. These are associated
with underlying \verb{stat_*} functions used for computation. Both are described
in their own manual pages.
}
\keyword{datasets}
\keyword{internal}
