% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_measures.R
\name{is_angle_in_range}
\alias{is_angle_in_range}
\title{Angular range check}
\usage{
is_angle_in_range(x, angular_range = c(0, 2 * pi))
}
\arguments{
\item{x}{Numeric vector with angles to be tested.}

\item{angular_range}{Numeric vector with only the limits of
the angular test range. The full circle range in radians, \code{angular_range = c(0, 2*pi)},
is chosen by default.}

\item{period}{A numeric value declaring the period of the circular data. The
default is \code{2 * pi}.}
}
\value{
A logical vector of same length as x with values \code{TRUE} if the
individual angle is within the test range or \code{FALSE} if not.
}
\description{
This function checks if angles are within a specified range. The angles \code{x}
and the test range \code{angular_range} need to be in the same units but can be
any circular revolution of the angles ( \code{a = a + period * n} ).
}
