% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_measures.R
\name{nice_angles}
\alias{nice_angles}
\title{Nice angles}
\usage{
nice_angles(x, period = 2 * pi, force_radians = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{period}{A numeric value declaring the period of the circular data. The
default is \code{2 * pi}.}

\item{force_radians}{A boolean value deciding if the conversion is to the
unit cycle. The default value is \code{TRUE}, converting the output on the unit
cycle \verb{[0, 2*pi[}.}
}
\value{
A numeric vector with circular data converted to the range
\verb{[0, period[}.
}
\description{
The function \code{nice_angles} converts cyclical data to the standard interval
\verb{[0, period[} or unit cycle \verb{[0, 2*pi[} to simplify calculations.
}
