% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional_speed.R
\name{vector_mean_speed}
\alias{vector_mean_speed}
\title{Vector mean velocity (Yamartino 1984)}
\usage{
vector_mean_speed(angle, period = 2 * pi, speed)
}
\arguments{
\item{angle}{Numeric vector with angles}

\item{period}{A numeric value declaring the period of the circular data. The
default is \code{2 * pi}.}

\item{speed}{Numeric vector with velocities used to weight the mean}
}
\value{
The average angle in radians or degrees.
}
\description{
Besides the mean speed \eqn{\bar{V}=\frac{1}{n}\sum_{i=1}^n V_i}, the vector-
mean speed can be calculated with the average speed in the x- and y-
direction, respectively:
\deqn{V_v=\sqrt{\bar{V_x}^2 + \bar{V_y}^2} = \sqrt{\bar{-\frac{1}{n}
\sum_{i=1}^n V_i\text{sin}(\theta_i)}^2 + \bar{-\frac{1}{n}\sum_{i=1}^n V_i
\text{cos}(\theta_i)}^2}}
}
\seealso{
Other directional_speed: 
\code{\link{angular_sd_yamartino}()},
\code{\link{directional_speed}},
\code{\link{weighted_angular_mean}()}

Other location_measures: 
\code{\link{angular_mean}()},
\code{\link{circular_location}},
\code{\link{weighted_angular_mean}()}
}
\concept{directional_speed}
\concept{location_measures}
