% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional_speed.R
\name{weighted_angular_mean}
\alias{weighted_angular_mean}
\title{Speed-weighted angular mean (Yamartino 1984)}
\usage{
weighted_angular_mean(angle, speed, period = 2 * pi)
}
\arguments{
\item{angle}{Numeric vector with angles}

\item{speed}{Numeric vector with velocities used to weight the mean}

\item{period}{A numeric value declaring the period of the circular data. The
default is \code{2 * pi}.}
}
\value{
The average angle in radians or degrees.
}
\description{
Angular data repeat every 2*pi radians which has to be taken into account
for calculating the angular means. For directional velocities, the speed
values can be used to weight the mean using this function based on the
speed-weighted angular mean (Yamartino 1984).
}
\details{
The mean wind direction weighted by the wind speed is calculated by
\deqn{\theta_v = \text{atan}(V_{xa},V_{ya})}
with the average velocity in the x (\eqn{V_{xa} = \frac{1}{n}\sum_{i=1}^n V_i
\cdot\text{sin}(\theta_i)}) and y (\eqn{V_{ya} = \frac{1}{n}\sum_{i=1}^n V_i
\cdot\text{cos}(\theta_i)}) directions, respectively.
}
\seealso{
Other directional_speed: 
\code{\link{angular_sd_yamartino}()},
\code{\link{directional_speed}}

Other location_measures: 
\code{\link{angular_mean}()},
\code{\link{circular_location}}
}
\concept{directional_speed}
\concept{location_measures}
